pyproject.toml
[tool.poetry]
name = "calc"
version = "0.1.0"
description = ""
authors = ["Artsiom Beida (Artemiusz Bejda)"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
pydantic = "^1.9.2"
uvicorn = "^0.18.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^5.0.4"
types-psycopg2 = "^2.9.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
